+++ kernel/src/Graphics/Graphics.h
@@ -3,17 +3,26 @@
 
 #include <stdint.h>
 
-typedef struct {
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+
+struct Position {
     uint64_t x;
     uint64_t y;
-} Position;
+};
 
-typedef struct {
+struct FrameBuffer {
     void* FrameBufferAddress;
     uint64_t FrameBufferWidth;
     uint64_t FrameBufferHeight;
     uint16_t bpp; // Bits per pixel
-} FrameBuffer;
+};
 
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif /* _X86_64_GRAPHICS_DEFINITIONS_H */